Issue #24858 allow dataStore to be a persistence unit reference
